chore: rearrange types and correct component's description
