+++ packages/fiori/src/Timeline.ts
@@ -7,16 +7,29 @@ import { getI18nBundle } from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import { isTabNext, isTabPrevious } from "@ui5/webcomponents-base/dist/Keys.js";
 import ItemNavigation from "@ui5/webcomponents-base/dist/delegate/ItemNavigation.js";
+import type { ITabbable } from "@ui5/webcomponents-base/dist/delegate/ItemNavigation.js";
 import NavigationMode from "@ui5/webcomponents-base/dist/types/NavigationMode.js";
 import { getEventMark } from "@ui5/webcomponents-base/dist/MarkedEvents.js";
 import { TIMELINE_ARIA_LABEL } from "./generated/i18n/i18n-defaults.js";
 import TimelineTemplate from "./generated/templates/TimelineTemplate.lit.js";
 import TimelineItem from "./TimelineItem.js";
-import type { ITimelineItem } from "./Interfaces.js";
 
 // Styles
 import TimelineCss from "./generated/themes/Timeline.css.js";
 import TimelineLayout from "./types/TimelineLayout.js";
+
+/**
+ * Interface for components that may be slotted inside <code>ui5-timeline</code> as items
+ *
+ * @public
+ */
+interface ITimelineItem extends UI5Element, ITabbable {
+    layout: `${TimelineLayout}`,
+    icon: string,
+    forcedLineWidth: string,
+    nameClickable: boolean,
+    focusLink: () => void,
+}
 
 const SHORT_LINE_WIDTH = "ShortLineWidth";
 const LARGE_LINE_WIDTH = "LargeLineWidth";
@@ -116,9 +129,9 @@ class Timeline extends UI5Element {
 		for (let i = 0; i < this.items.length; i++) {
 			this.items[i].layout = this.layout;
 			if (this.items[i + 1] && !!this.items[i + 1].icon) {
-				this.items[i]._lineWidth = SHORT_LINE_WIDTH;
+				this.items[i].forcedLineWidth = SHORT_LINE_WIDTH;
 			} else if (this.items[i].icon && this.items[i + 1] && !this.items[i + 1].icon) {
-				this.items[i]._lineWidth = LARGE_LINE_WIDTH;
+				this.items[i].forcedLineWidth = LARGE_LINE_WIDTH;
 			}
 		}
 	}
@@ -156,3 +169,6 @@ class Timeline extends UI5Element {
 Timeline.define();
 
 export default Timeline;
+export type {
+	ITimelineItem,
+};
