+++ packages/fiori/src/UploadCollection.ts
@@ -28,7 +28,6 @@ import {
 	detachBodyDnDHandler,
 	draggingFiles,
 } from "./upload-utils/UploadCollectionBodyDnD.js";
-import type IUploadCollectionItem from "./UploadCollectionItem.js";
 import type { DnDEventListener, DnDEventListenerParam } from "./upload-utils/UploadCollectionBodyDnD.js";
 import UploadCollectionDnDOverlayMode from "./types/UploadCollectionDnDMode.js";
 
@@ -37,6 +36,13 @@ import UploadCollectionTemplate from "./generated/templates/UploadCollectionTemp
 
 // Styles
 import UploadCollectionCss from "./generated/themes/UploadCollection.css.js";
+
+/**
+ * Interface for components that may be slotted inside <code>ui5-upload-collection</code> as items
+ *
+ * @public
+ */
+interface IUploadCollectionItem extends HTMLElement { }
 
 type UploadCollectionSelectionChangeEventDetail = {
 	selectedItems: Array<IUploadCollectionItem>,
@@ -344,6 +350,7 @@ UploadCollection.define();
 
 export default UploadCollection;
 export type {
+	IUploadCollectionItem,
 	UploadCollectionItemDeleteEventDetail,
 	UploadCollectionSelectionChangeEventDetail,
 };
