+++ packages/main/src/Breadcrumbs.ts
@@ -210,11 +210,11 @@ class Breadcrumbs extends UI5Element {
 		this._labelFocusAdaptor = {
 			id: `${this._id}-labelWrapper`,
 			getlabelWrapper: this.getCurrentLocationLabelWrapper.bind(this),
-			set _tabIndex(value: string) {
+			set forcedTabIndex(value: string) {
 				const wrapper = this.getlabelWrapper();
 				wrapper && wrapper.setAttribute("tabindex", value);
 			},
-			get _tabIndex() {
+			get forcedTabIndex() {
 				const wrapper = this.getlabelWrapper();
 				return wrapper?.getAttribute("tabindex") || "";
 			},
@@ -373,8 +373,8 @@ class Breadcrumbs extends UI5Element {
 		// if the last focused link has done into the overflow =>
 		// ensure the first visible link is focusable
 		const focusableItems = this._getFocusableItems();
-		if (!focusableItems.some(x => x._tabIndex === "0")) {
-			this._itemNavigation.setCurrentItem(focusableItems[0]);
+		if (!focusableItems.some(x => x.forcedTabIndex === "0")) {
+			this.forcedTabIndex.setCurrentItem(focusableItems[0]);
 		}
 	}
 
@@ -535,7 +535,7 @@ class Breadcrumbs extends UI5Element {
 	}
 
 	get _isDropdownArrowFocused() {
-		return this._dropdownArrowLink._tabIndex === "0";
+		return this._dropdownArrowLink.forcedTabIndex === "0";
 	}
 
 	get _isCurrentLocationLabelFocused() {
