+++ packages/main/src/ColorPalettePopover.ts
@@ -22,9 +22,8 @@ import Button from "./Button.js";
 import Title from "./Title.js";
 import ResponsivePopover from "./ResponsivePopover.js";
 import ColorPalette from "./ColorPalette.js";
-import type { ColorPaletteItemClickEventDetail } from "./ColorPalette.js";
+import type { ColorPaletteItemClickEventDetail, IColorPaletteItem } from "./ColorPalette.js";
 import type ColorPaletteItem from "./ColorPaletteItem.js";
-import { IColorPaletteItem } from "./Interfaces.js";
 
 type ColorPalettePopoverItemClickEventDetail = ColorPaletteItemClickEventDetail;
