+++ packages/main/src/Option.ts
@@ -2,7 +2,7 @@ import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
-import type { IOption } from "./Interfaces.js";
+import type { IOption } from "./Select.js";
 /**
  * @class
  *
@@ -14,6 +14,7 @@ import type { IOption } from "./Interfaces.js";
  * @extends UI5Element
  * @implements {IOption}
  * @public
+ * @abstract
  */
 @customElement("ui5-option")
 class Option extends UI5Element implements IOption {
