+++ packages/main/src/SegmentedButtonItem.ts
@@ -2,7 +2,7 @@ import customElement from "@ui5/webcomponents-base/dist/decorators/customElement
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 
 import Integer from "@ui5/webcomponents-base/dist/types/Integer.js";
-import type { ISegmentedButtonItem } from "./Interfaces.js";
+import type { ISegmentedButtonItem } from "./SegmentedButton.js";
 import SegmentedButtonItemTemplate from "./generated/templates/SegmentedButtonItemTemplate.lit.js";
 
 import ToggleButton from "./ToggleButton.js";
@@ -30,7 +30,6 @@ import { SEGMENTEDBUTTONITEM_ARIA_DESCRIPTION } from "./generated/i18n/i18n-defa
  * @constructor
  * @extends ToggleButton
  * @implements { ISegmentedButtonItem }
- * @abstract
  * @public
  */
 @customElement({
