+++ packages/main/src/TableRow.ts
@@ -154,7 +154,7 @@ class TableRow extends UI5Element implements ITableRow {
 	_columnsInfo!: Array<TableColumnInfo>;
 
 	@property({ defaultValue: "-1" })
-	_tabIndex!: string;
+	forcedTabIndex!: string;
 
 	@property({ type: Boolean })
 	_busy!: boolean;
