+++ packages/main/src/TimeSelectionClocks.ts
@@ -47,7 +47,6 @@ import TimeSelectionClocksCss from "./generated/themes/TimeSelectionClocks.css.j
  *
  * @constructor
  * @extends TimePickerInternals
- * @abstract
  * @since 1.15.0
  * @private
  */
