+++ packages/main/src/Toolbar.ts
@@ -468,14 +468,14 @@ class Toolbar extends UI5Element {
 	onOverflowPopoverClosed() {
 		this.popoverOpen = false;
 		if (this.overflowButtonDOM) {
-			this.overflowButtonDOM.accessibilityAttributes.expanded = "false";
+			this.overflowButtonDOM.accessibilityAttributes.expanded = false;
 		}
 	}
 
 	onOverflowPopoverOpened() {
 		this.popoverOpen = true;
 		if (this.overflowButtonDOM) {
-			this.overflowButtonDOM.accessibilityAttributes.expanded = "true";
+			this.overflowButtonDOM.accessibilityAttributes.expanded = true;
 		}
 	}
