+++ packages/main/src/ToolbarButton.ts
@@ -3,6 +3,7 @@ import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import event from "@ui5/webcomponents-base/dist/decorators/event.js";
 import CSSSize from "@ui5/webcomponents-base/dist/types/CSSSize.js";
 import Button from "./Button.js";
+import type { AccessibilityAttributes as ButtonAccessibilityAttributes } from "./Button.js";
 import ButtonDesign from "./types/ButtonDesign.js";
 
 import ToolbarItem from "./ToolbarItem.js";
@@ -13,6 +14,8 @@ import ToolbarPopoverButtonTemplate from "./generated/templates/ToolbarPopoverBu
 import ToolbarButtonPopoverCss from "./generated/themes/ToolbarButtonPopover.css.js";
 
 import { registerToolbarItem } from "./ToolbarRegistry.js";
+
+type AccessibilityAttributes = ButtonAccessibilityAttributes;
 
 /**
  * @class
@@ -141,11 +144,11 @@ class ToolbarButton extends ToolbarItem {
 	 * 		</li>
 	 * 		<li><code>hasPopup</code>: Indicates the availability and type of interactive popup element, such as menu or dialog, that can be triggered by the button. Accepts the following string values:
 	 * 			<ul>
-	 *				<li><code>Dialog</code></li>
-	 *				<li><code>Grid</code></li>
-	 *				<li><code>ListBox</code></li>
-	 *				<li><code>Menu</code></li>
-	 *				<li><code>Tree</code></li>
+	 *				<li><code>dialog</code></li>
+	 *				<li><code>grid</code></li>
+	 *				<li><code>listbox</code></li>
+	 *				<li><code>menu</code></li>
+	 *				<li><code>tree</code></li>
 	 * 			</ul>
 	 * 		</li>
 	 * 		<li><code>controls</code>: Identifies the element (or elements) whose contents or presence are controlled by the button element. Accepts a string value.</li>
@@ -154,7 +157,7 @@ class ToolbarButton extends ToolbarItem {
 	 * @public
 	 */
 	@property({ type: Object })
-	accessibilityAttributes!: { expanded: "true" | "false", hasPopup: "Dialog" | "Grid" | "ListBox" | "Menu" | "Tree", controls: string};
+	accessibilityAttributes!: AccessibilityAttributes;
 
 	/**
 	 * Button text
@@ -211,3 +214,7 @@ registerToolbarItem(ToolbarButton);
 ToolbarButton.define();
 
 export default ToolbarButton;
+
+export type {
+	AccessibilityAttributes,
+};
